% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_shifted_data}
\alias{create_shifted_data}
\title{Shift Exposures}
\usage{
create_shifted_data(data, exposure, delta, lower_bound, upper_bound)
}
\arguments{
\item{data}{The data}

\item{exposure}{The exposure}

\item{delta}{How much to shift}

\item{lower_bound}{lowest level of exposure}

\item{upper_bound}{highest level of exposure}
}
\description{
Shifts Exposures by delta
}
