% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_density_ratio}
\alias{estimate_density_ratio}
\title{Estimate Density Ratio via Classification}
\usage{
estimate_density_ratio(at, av, delta, var, covars, classifier)
}
\arguments{
\item{at}{Data frame containing the training data.}

\item{av}{Data frame containing the validation data.}

\item{delta}{Numeric value specifying the shift to be applied to the exposure variable.}

\item{var}{Character string specifying the name of the exposure variable to be shifted.}

\item{covars}{Character vector specifying the names of the covariate columns in the data frames.}

\item{classifier}{An \code{sl3} learner object used for classification.}
}
\value{
A list containing two elements: \code{Hn_at} and \code{Hn_av}, which are data tables of density ratios for the
training and validation datasets, respectively. Each data table has two columns: \code{noshift} and \code{shift},
representing the density ratios for unshifted and shifted exposures.
}
\description{
This function estimates the density ratio of shifted versus unshifted exposures using a classification approach.
It utilizes the \code{sl3} package to train a Super Learner on the intervention indicator and then predicts the density ratios
for both the training and validation datasets.
}
\examples{
# Example usage:
training_data <- data.frame(A = rnorm(100), W1 = rbinom(100, 1, 0.5), W2 = rbinom(100, 1, 0.5))
validation_data <- data.frame(A = rnorm(50), W1 = rbinom(50, 1, 0.5), W2 = rbinom(50, 1, 0.5))
delta <- -0.5
exposure_variable <- "A"
covariates <- c("W1", "W2")
classifier <- sl3::Lrnr_glm$new()
result <- estimate_density_ratio(training_data, validation_data, delta, exposure_variable, covariates, classifier)
training_density_ratios <- result$Hn_at
validation_density_ratios <- result$Hn_av

}
