% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterXshift.R
\name{InterXshift}
\alias{InterXshift}
\title{InterXshift}
\usage{
InterXshift(
  w,
  a,
  y,
  deltas,
  estimator = "tmle",
  fluctuation = "standard",
  var_sets = NULL,
  pi_learner = NULL,
  mu_learner = NULL,
  g_learner = NULL,
  e_learner = NULL,
  zeta_learner = NULL,
  n_folds = 2,
  outcome_type = "continuous",
  parallel = TRUE,
  parallel_type = "multi_session",
  num_cores = 2,
  seed = seed,
  hn_trunc_thresh = 10,
  adaptive_delta = FALSE,
  discover_only = FALSE,
  top_n = 2
)
}
\arguments{
\item{w}{A \code{matrix}, \code{data.frame}, or similar containing a set of
baseline covariates. These variables are measured before exposures.}

\item{a}{\code{matrix}, \code{data.frame}, or similar containing individual or
multiple exposures.}

\item{y}{\code{numeric} vector of observed outcomes.}

\item{deltas}{A \code{numeric} value indicating the shift in exposures to
define the target parameter, with respect to the scale of the exposures (A). If adaptive_delta
is true, these values will be reduced.}

\item{estimator}{The type of estimator to fit: \code{"tmle"} for targeted
maximum likelihood estimation, or \code{"onestep"} for a one-step estimator.}

\item{fluctuation}{Method used in the targeting step for TML estimation: "standard" or "weighted".
This determines where to place the auxiliary covariate in the logistic tilting regression.}

\item{var_sets}{A list specifying variable sets for deterministic InterXshift usage.
Example: var_sets <- c("A_1", "A_1-Z_2") where the analyst provides variable sets
for exposures, exposure-mediator, or exposure-covariate relationships.}

\item{pi_learner}{Learners for fitting Super Learner ensembles to densities via \pkg{sl3}.}

\item{mu_learner}{Learners for fitting Super Learner ensembles to the outcome model via \pkg{sl3}.}

\item{g_learner}{Learners for fitting Super Learner ensembles to the g-mechanism
g(A|W) (a probability estimator, not a density estimator) for mediation via \pkg{sl3}.}

\item{e_learner}{Learners for fitting Super Learner ensembles to the e-mechanism
g(A|Z,W) (a probability estimator, not a density estimator) for mediation via \pkg{sl3}.}

\item{zeta_learner}{Learners for fitting Super Learner ensembles to the outcome model via \pkg{sl3}..}

\item{n_folds}{Number of folds to use in cross-validation, default is 2.}

\item{outcome_type}{Data type of the outcome, default is "continuous".}

\item{parallel}{Whether to parallelize across cores (default: TRUE).}

\item{parallel_type}{Type of parallelization to use if parallel is TRUE:
"multi_session" (default), "multicore", or "sequential".}

\item{num_cores}{Number of CPU cores to use in parallelization (default: 2).}

\item{seed}{\code{numeric} seed value to be passed to all functions.}

\item{hn_trunc_thresh}{Truncation level for the clever covariate (default: 10).}

\item{adaptive_delta}{If TRUE, reduces the user-specified delta until
the Hn calculated for a shift does not have any observation greater
than hn_trunc_thresh (default: FALSE).}

\item{z}{\code{matrix}, \code{data.frame}, or similar containing individual or
multiple mediators (optional).}
}
\value{
An S3 object of class \code{SuperNOVA} containing the results of the
procedure to compute a TML or one-step estimate of the counterfactual mean
under a modified treatment policy that shifts a continuous-valued exposure
by a scalar amount \code{delta}. These exposures are data-adaptively
identified using the CV-TMLE procedure.
}
\description{
Under a fixed shift to exposures identify using g-computation the joint shift of
pairwise exposures in a mixed exposure compared to the additive individual shifts. Positive values indicate
synergy and negative antagonism, get the top synergy and antagonism results and use CV-TMLE to efficiently
estimate the interaction target parameter.
}
